requirements.txt
pandas
numpy
pretty_midi
torch>=2.0.0
torchvision
torchaudio
librosa
sounddevice
matplotlib
scikit-learn
openai
PyYAML
transformers>=4.34.0
datasets>=2.14.0
peft>=0.6.0
accelerate>=0.21.0
bitsandbytes>=0.41.0
trl>=0.7.2
scipy>=1.10.0
sentencepiece>=0.1.99
protobuf>=3.20.0
tensorboard>=2.13.0
wandb>=0.15.0
huggingface_hub>=0.17.0
